Atlas_AI/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
pgvector==0.2.4
openai==1.3.0
python-dotenv==1.0.0
markdown==3.5.1
